Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/workers/WorkerGlobalScope_ErrorEvent_filename-expected.txt

CONSOLE MESSAGE: Error: Error Message

PASS  WorkerGlobalScope onerror event handler argument: location
